refactor: move task-queue file to `static/js/`
