+++ static/js/workers/git.worker.js
@@ -1,6 +1,6 @@
 import { isImageExtension } from '../helpers/image.js';
 import { Octokit } from '../vendor/octokit-core-6.1.3.min.js';
-import TaskQueue from '../helpers/task-queue.js';
+import TaskQueue from '../task-queue.js';
 
 const GITHUB_REPO_URL_PATTERN = /^https:\/\/github.com\/([\w-]+)\/([\w-]+)(?:\.git)?/;
